+++ apps/evalExportTrajectories.cpp
@@ -19,7 +19,7 @@
 
 #include "AppUtils.h"
 #include "EKFNEESEvaluator.h"
-#include "PIMs.h"
+#include "ResultsAdapters.h"
 #include "ResultsWriter.h"
 
 using namespace gtsam;
@@ -118,121 +118,6 @@ DatasetFilter selectFilter(const string& datasetType) {
   return DatasetFilters::all;
 }
 
-TrajectorySampleRow makeTrajectoryRow(
-    const ResultsWriter& writer, const string& datasetName,
-    const string& configLabel, const EKFNEESEvaluator::RunArtifacts& artifacts,
-    const EKFNEESEvaluator::TrajectorySample& sample) {
-  TrajectorySampleRow row;
-  row.runId = writer.runId();
-  row.appName = writer.appName();
-  row.dataset = datasetName;
-  row.method = "gal3_imu_ekf";
-  row.configLabel = configLabel;
-  row.intervalSeconds = artifacts.preintegrationTime;
-  row.samplesPerWindow = artifacts.samplesPerWindow;
-  row.timestamp = sample.groundTruth.timestamp;
-  row.gtX = sample.groundTruth.position.x();
-  row.gtY = sample.groundTruth.position.y();
-  row.gtZ = sample.groundTruth.position.z();
-  row.gtVx = sample.groundTruth.velocity.x();
-  row.gtVy = sample.groundTruth.velocity.y();
-  row.gtVz = sample.groundTruth.velocity.z();
-  row.gtRoll = sample.groundTruth.rpy.x();
-  row.gtPitch = sample.groundTruth.rpy.y();
-  row.gtYaw = sample.groundTruth.rpy.z();
-  row.predX = sample.predicted.position.x();
-  row.predY = sample.predicted.position.y();
-  row.predZ = sample.predicted.position.z();
-  row.predVx = sample.predicted.velocity.x();
-  row.predVy = sample.predicted.velocity.y();
-  row.predVz = sample.predicted.velocity.z();
-  row.predRoll = sample.predicted.rpy.x();
-  row.predPitch = sample.predicted.rpy.y();
-  row.predYaw = sample.predicted.rpy.z();
-  row.errRotX = sample.error(0);
-  row.errRotY = sample.error(1);
-  row.errRotZ = sample.error(2);
-  row.errPosX = sample.error(3);
-  row.errPosY = sample.error(4);
-  row.errPosZ = sample.error(5);
-  row.errVelX = sample.error(6);
-  row.errVelY = sample.error(7);
-  row.errVelZ = sample.error(8);
-  row.rotPredSigma = covarianceBlockSigma(sample.predicted.covariance, 0);
-  row.posPredSigma = covarianceBlockSigma(sample.predicted.covariance, 3);
-  row.velPredSigma = covarianceBlockSigma(sample.predicted.covariance, 6);
-  row.covariance = sample.predicted.covariance;
-  return row;
-}
-
-WindowSummaryRow makeSummaryRow(const ResultsWriter& writer,
-                                const string& datasetName,
-                                const string& configLabel,
-                                const EKFNEESEvaluator::RunArtifacts& artifacts,
-                                const WindowResultSummary& summary) {
-  WindowSummaryRow row;
-  row.runId = writer.runId();
-  row.appName = writer.appName();
-  row.dataset = datasetName;
-  row.method = "gal3_imu_ekf";
-  row.configLabel = configLabel;
-  row.intervalSeconds = artifacts.preintegrationTime;
-  row.samplesPerWindow = artifacts.samplesPerWindow;
-  row.quadratureNodes = 0;
-  row.sampleCount = summary.sampleCount;
-  row.normalizedNeesMean = summary.normalizedNeesMean;
-  row.normalizedNeesMedian = summary.normalizedNeesMedian;
-  row.normalizedNeesP95 = summary.normalizedNeesP95;
-  row.normalizedNeesVariance = summary.normalizedNeesVariance;
-  row.rotErrorMedian = summary.rotErrorMedian;
-  row.rotPredSigmaMedian = summary.rotPredSigmaMedian;
-  row.posErrorMedian = summary.posErrorMedian;
-  row.posPredSigmaMedian = summary.posPredSigmaMedian;
-  row.velErrorMedian = summary.velErrorMedian;
-  row.velPredSigmaMedian = summary.velPredSigmaMedian;
-  return row;
-}
-
-void writeArtifacts(ResultsWriter* writer, const string& datasetName,
-                    const string& configLabel,
-                    const EKFNEESEvaluator::RunArtifacts& artifacts) {
-  vector<WindowResult> results;
-  results.reserve(artifacts.windowEvaluations.size());
-  for (const auto& evaluation : artifacts.windowEvaluations) {
-    WindowMetricRow row;
-    row.runId = writer->runId();
-    row.appName = writer->appName();
-    row.dataset = datasetName;
-    row.method = "gal3_imu_ekf";
-    row.configLabel = configLabel;
-    row.intervalSeconds = artifacts.preintegrationTime;
-    row.samplesPerWindow = artifacts.samplesPerWindow;
-    row.quadratureNodes = 0;
-    row.windowIndex = evaluation.windowIndex;
-    row.windowStartSample = evaluation.startSample;
-    row.windowEndSample = evaluation.endSample;
-    row.windowStartTime = evaluation.startTime;
-    row.windowEndTime = evaluation.endTime;
-    row.normalizedNees = evaluation.metrics.normalizedNees;
-    row.rotErrorNorm = evaluation.metrics.rotErrorNorm;
-    row.rotPredSigma = evaluation.metrics.rotPredSigma;
-    row.posErrorNorm = evaluation.metrics.posErrorNorm;
-    row.posPredSigma = evaluation.metrics.posPredSigma;
-    row.velErrorNorm = evaluation.metrics.velErrorNorm;
-    row.velPredSigma = evaluation.metrics.velPredSigma;
-    writer->writeWindowMetric(row);
-    results.push_back(evaluation.metrics);
-  }
-  writer->writeWindowSummary(makeSummaryRow(*writer, datasetName, configLabel,
-                                            artifacts,
-                                            summarizeWindowResults(results)));
-
-  for (const auto& sample : artifacts.trajectorySamples) {
-    writer->writeTrajectorySample(makeTrajectoryRow(
-        *writer, datasetName, configLabel, artifacts, sample));
-  }
-}
-
 }  // namespace
 
 int main(int argc, char* argv[]) {
@@ -246,29 +131,26 @@ int main(int argc, char* argv[]) {
     }
 
     ResultsWriter writer(argv[0], options.outputRoot);
-    writer.writeRunMetadata({writer.runId(), writer.appName(),
-                             writer.timestampUtc(),
-                             joinCommandLineArguments(argc, argv),
-                             writer.outputRoot().string(), ""});
-    for (const auto& [datasetName, datasetPath] : datasets) {
-      writer.writeDataset({writer.runId(), writer.appName(), datasetName,
-                           datasetPath, options.datasetType});
-    }
+    writeCanonicalRunMetadata(&writer, argc, argv);
 
     for (const auto& [datasetName, datasetPath] : datasets) {
+      writer.writeDataset(makeDatasetRow(writer, datasetName, datasetPath,
+                                         options.datasetType));
       Dataset dataset(datasetPath);
       EKFNEESEvaluator evaluator(dataset);
 
-      writeArtifacts(&writer, datasetName, options.best.label,
-                     evaluator.computeGal3ImuEKFArtifacts(
-                         kIntervalSeconds,
-                         dataset.configureImuParams(options.best.alphaGyro,
-                                                    options.best.alphaAcc)));
-      writeArtifacts(&writer, datasetName, options.worst.label,
-                     evaluator.computeGal3ImuEKFArtifacts(
-                         kIntervalSeconds,
-                         dataset.configureImuParams(options.worst.alphaGyro,
-                                                    options.worst.alphaAcc)));
+      writeEkfArtifacts(&writer, datasetName, "gal3_imu_ekf",
+                        options.best.label,
+                        evaluator.computeGal3ImuEKFArtifacts(
+                            kIntervalSeconds,
+                            dataset.configureImuParams(options.best.alphaGyro,
+                                                       options.best.alphaAcc)));
+      writeEkfArtifacts(
+          &writer, datasetName, "gal3_imu_ekf", options.worst.label,
+          evaluator.computeGal3ImuEKFArtifacts(
+              kIntervalSeconds,
+              dataset.configureImuParams(options.worst.alphaGyro,
+                                         options.worst.alphaAcc)));
     }
 
     cout << "Results written to " << writer.runDirectory() << "\n";
