+++ apps/evalReducedNeesWithPriorCovariance.cpp
@@ -19,14 +19,13 @@
 
 #include <cmath>
 #include <iostream>
-#include <optional>
 #include <string>
 #include <vector>
 
 #include "AppUtils.h"
 #include "PIMs.h"
+#include "ResultsAdapters.h"
 #include "ResultsWriter.h"
-#include "Window.h"
 
 using namespace gtsam;
 using namespace std;
@@ -48,15 +47,6 @@ struct AppOptions {
   size_t maxIntervals = 0;
 };
 
-struct WindowEvaluationRecord {
-  size_t windowIndex = 0;
-  size_t startSample = 0;
-  size_t endSample = 0;
-  double startTime = 0.0;
-  double endTime = 0.0;
-  WindowResult metrics;
-};
-
 Matrix9 initialNavCovariance() {
   return Matrix9::Identity() * kInitialStateCovariance;
 }
@@ -71,29 +61,6 @@ shared_ptr<PreintegrationParams> createPreintegrationParams() {
   params->gyroscopeCovariance = I_3x3 * kSigmaGyro * kSigmaGyro;
   params->integrationCovariance = I_3x3 * 1e-8;
   return params;
-}
-
-template <class PIMType>
-vector<WindowEvaluationRecord> collectWindowEvaluations(
-    const vector<Window>& windows, const shared_ptr<PreintegrationParams>& params,
-    size_t quadratureOrder = 0) {
-  vector<WindowEvaluationRecord> evaluations;
-  evaluations.reserve(windows.size());
-  const InitialCovarianceOptions initialCovariance{initialNavCovariance(),
-                                                   initialBiasCovariance()};
-  for (size_t windowIndex = 0; windowIndex < windows.size(); ++windowIndex) {
-    const auto& window = windows[windowIndex];
-    const auto result =
-        evaluateWindow<PIMType>(window, params, initialCovariance,
-                                quadratureOrder);
-    if (!result) {
-      continue;
-    }
-    evaluations.push_back({windowIndex, window.start, window.end,
-                           window.initialTruth().timestamp,
-                           window.terminalTruth().timestamp, *result});
-  }
-  return evaluations;
 }
 
 void printUsage(const char* programName) {
@@ -113,73 +80,6 @@ AppOptions parseAppArguments(const vector<string>& arguments,
   return {parseMaxIntervalsArgument(arguments)};
 }
 
-WindowSummaryRow makeSummaryRow(const ResultsWriter& writer,
-                                const string& datasetName,
-                                const string& method,
-                                double intervalSeconds,
-                                size_t samplesPerWindow,
-                                size_t quadratureNodes,
-                                const WindowResultSummary& summary) {
-  WindowSummaryRow row;
-  row.runId = writer.runId();
-  row.appName = writer.appName();
-  row.dataset = datasetName;
-  row.method = method;
-  row.configLabel = kConfigLabel;
-  row.intervalSeconds = intervalSeconds;
-  row.samplesPerWindow = samplesPerWindow;
-  row.quadratureNodes = quadratureNodes;
-  row.sampleCount = summary.sampleCount;
-  row.normalizedNeesMean = summary.normalizedNeesMean;
-  row.normalizedNeesMedian = summary.normalizedNeesMedian;
-  row.normalizedNeesP95 = summary.normalizedNeesP95;
-  row.normalizedNeesVariance = summary.normalizedNeesVariance;
-  row.rotErrorMedian = summary.rotErrorMedian;
-  row.rotPredSigmaMedian = summary.rotPredSigmaMedian;
-  row.posErrorMedian = summary.posErrorMedian;
-  row.posPredSigmaMedian = summary.posPredSigmaMedian;
-  row.velErrorMedian = summary.velErrorMedian;
-  row.velPredSigmaMedian = summary.velPredSigmaMedian;
-  return row;
-}
-
-void writeWindowEvaluations(const ResultsWriter& writer, ResultsWriter* sink,
-                            const string& datasetName, const string& method,
-                            double intervalSeconds, size_t samplesPerWindow,
-                            size_t quadratureNodes,
-                            const vector<WindowEvaluationRecord>& evaluations) {
-  vector<WindowResult> results;
-  results.reserve(evaluations.size());
-  for (const auto& evaluation : evaluations) {
-    WindowMetricRow row;
-    row.runId = writer.runId();
-    row.appName = writer.appName();
-    row.dataset = datasetName;
-    row.method = method;
-    row.configLabel = kConfigLabel;
-    row.intervalSeconds = intervalSeconds;
-    row.samplesPerWindow = samplesPerWindow;
-    row.quadratureNodes = quadratureNodes;
-    row.windowIndex = evaluation.windowIndex;
-    row.windowStartSample = evaluation.startSample;
-    row.windowEndSample = evaluation.endSample;
-    row.windowStartTime = evaluation.startTime;
-    row.windowEndTime = evaluation.endTime;
-    row.normalizedNees = evaluation.metrics.normalizedNees;
-    row.rotErrorNorm = evaluation.metrics.rotErrorNorm;
-    row.rotPredSigma = evaluation.metrics.rotPredSigma;
-    row.posErrorNorm = evaluation.metrics.posErrorNorm;
-    row.posPredSigma = evaluation.metrics.posPredSigma;
-    row.velErrorNorm = evaluation.metrics.velErrorNorm;
-    row.velPredSigma = evaluation.metrics.velPredSigma;
-    sink->writeWindowMetric(row);
-    results.push_back(evaluation.metrics);
-  }
-  sink->writeWindowSummary(makeSummaryRow(
-      writer, datasetName, method, intervalSeconds, samplesPerWindow,
-      quadratureNodes, summarizeWindowResults(results)));
-}
-
 struct RunForDataset {
   RunForDataset(const AppOptions& options, ResultsWriter* writer)
       : intervals(selectIntervals(defaultQuadratureIntervals(),
@@ -188,8 +88,11 @@ struct RunForDataset {
 
   vector<double> intervals;
   ResultsWriter* writer = nullptr;
+  string datasetGroup = "all";
 
   void operator()(const string& datasetName, const Dataset& dataset) {
+    writer->writeDataset(
+        makeDatasetRow(*writer, datasetName, dataset, datasetGroup));
     for (double intervalSeconds : intervals) {
       const size_t samplesPerWindow = dataset.stepsForInterval(intervalSeconds);
       const size_t quadratureNodes = max<size_t>(
@@ -198,19 +101,24 @@ struct RunForDataset {
       const auto windows = dataset.completeWindows(samplesPerWindow);
       const auto params = createPreintegrationParams();
 
-      writeWindowEvaluations(
-          *writer, writer, datasetName, "quadrature", intervalSeconds,
+      const InitialCovarianceOptions initialCovariance{initialNavCovariance(),
+                                                       initialBiasCovariance()};
+      writeWindowRows(
+          writer, datasetName, "quadrature", kConfigLabel, intervalSeconds,
           samplesPerWindow, quadratureNodes,
           collectWindowEvaluations<PIMQuadrature>(windows, params,
+                                                  initialCovariance,
                                                   quadratureNodes));
-      writeWindowEvaluations(
-          *writer, writer, datasetName, "manifold", intervalSeconds,
+      writeWindowRows(
+          writer, datasetName, "manifold", kConfigLabel, intervalSeconds,
           samplesPerWindow, 0,
-          collectWindowEvaluations<PIMManifold>(windows, params));
-      writeWindowEvaluations(
-          *writer, writer, datasetName, "tangent", intervalSeconds,
+          collectWindowEvaluations<PIMManifold>(windows, params,
+                                                initialCovariance));
+      writeWindowRows(
+          writer, datasetName, "tangent", kConfigLabel, intervalSeconds,
           samplesPerWindow, 0,
-          collectWindowEvaluations<PIMTangent>(windows, params));
+          collectWindowEvaluations<PIMTangent>(windows, params,
+                                               initialCovariance));
     }
   }
 };
@@ -228,16 +136,11 @@ int main(int argc, char* argv[]) {
       parseAppArguments(datasetCli.remainingArgs, argv[0]);
 
   ResultsWriter writer(argv[0], datasetCli.options.outputRoot);
-  writer.writeRunMetadata(
-      {writer.runId(), writer.appName(), writer.timestampUtc(),
-       joinCommandLineArguments(argc, argv), writer.outputRoot().string(), ""});
+  writeCanonicalRunMetadata(&writer, argc, argv);
   const string datasetGroup = datasetGroupLabel(datasetCli);
-  for (const auto& [datasetName, datasetPath] : datasetCli.datasets) {
-    writer.writeDataset({writer.runId(), writer.appName(), datasetName,
-                         datasetPath, datasetGroup});
-  }
 
   RunForDataset runner(appOptions, &writer);
+  runner.datasetGroup = datasetGroup;
   const int status = runForDatasets(datasetCli, runner);
   if (status != 0) {
     return status;
